RUN-238 Refactoring of components with Compose functions
refactor AboutAppFragment
