+++ app/src/main/java/com/tnco/runar/ui/widgets/TopBarIcon.kt
@@ -1,0 +1,28 @@
+package com.tnco.runar.ui.widgets
+
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
+import com.tnco.runar.R
+
+@Composable
+fun TopBarIcon(navController: NavController) {
+    IconButton(onClick = { navController.popBackStack() }) {
+        Icon(
+            painter = painterResource(id = R.drawable.ic_library_back_arrow_2),
+            tint = colorResource(id = R.color.library_top_bar_fav),
+            contentDescription = "arrow"
+        )
+    }
+}
+
+@Preview
+@Composable
+fun TopBarIconPreview() {
+    TopBarIcon(navController = rememberNavController())
+}
